added ragam pitch converter
The ragam pitch converter attempts to merge the worlds of Carnatic music and Western music a little better! It asks the user to enter a shruthi (tonic pitch), arohanam (notes of the ascending scale), and avarohanam (notes of the descending scale). It outputs the ascending and descending scale in western music piano notes. I have used the Mingus Python 3 libraries to help map Carnatic notes to western music pitches. 
This tool could be helpful to Carnatic music teachers who want to use the aid of a piano to teach different ragams (musical scales) to their students.
Future goals for this project: I want to add an audio component to this project: a piano/synth widget that plays the ragam in the pitch specified by the user. At the bottom of the Jupyter notebook, I have shown a glimpse of which other libraries I can use to achieve this.